+++ src/cpx_io/cpx_system/cpx_e.py
@@ -168,7 +168,7 @@ class CpxE(CpxBase):
         return module
 
 
-class _CpxEModule(CpxBase):
+class CpxEModule(CpxBase):
     """Base class for cpx-e modules"""
 
     def __init__(self, name=None):
@@ -191,7 +191,7 @@ class _CpxEModule(CpxBase):
         self.position = position
 
 
-class CpxEEp(_CpxEModule):
+class CpxEEp(CpxEModule):
     """Class for CPX-E-EP module"""
 
     def _initialize(self, *args):
@@ -204,7 +204,7 @@ class CpxEEp(_CpxEModule):
         self.base._next_input_register = self.input_register + 3
 
 
-class CpxE8Do(_CpxEModule):
+class CpxE8Do(CpxEModule):
     """Class for CPX-E-8DO module"""
 
     def __getitem__(self, key):
@@ -334,7 +334,7 @@ class CpxE8Do(_CpxEModule):
         self.base.write_function_number(function_number, value_to_write)
 
 
-class CpxE16Di(_CpxEModule):
+class CpxE16Di(CpxEModule):
     """Class for CPX-E-16DI module"""
 
     def __getitem__(self, key):
@@ -449,7 +449,7 @@ class CpxE16Di(_CpxEModule):
         self.base.write_function_number(function_number, value_to_write)
 
 
-class CpxE4AiUI(_CpxEModule):
+class CpxE4AiUI(CpxEModule):
     """Class for CPX-E-4AI-UI module"""
 
     def __getitem__(self, key):
@@ -717,7 +717,7 @@ class CpxE4AiUI(_CpxEModule):
     # TODO: add more functions CPX-E-_AI-U-I_description_2020-01a_8126669g1.pdf chapter 3.3 ff.
 
 
-class CpxE4AoUI(_CpxEModule):
+class CpxE4AoUI(CpxEModule):
     """Class for CPX-E-4AO-UI module"""
 
     def __getitem__(self, key):
@@ -915,15 +915,307 @@ class CpxE4AoUI(_CpxEModule):
     # TODO: add more functions CPX-E-_AO-U-I_description_2020-01a_8126651g1.pdf chapter 3.3 ff.
 
 
-class CpxE4Iol(_CpxEModule):
+class CpxE4Iol(CpxEModule):
     """Class for CPX-E-4IOL io-link master module"""
 
-    # TODO: Add IO-Link module
-    def __init__(self):
-        raise NotImplementedError("The module CPX-E-4IOL has not yet been implemented")
+    def __getitem__(self, key):
+        return self.read_channel(key)
+
+    def __setitem__(self, key, value):
+        self.write_channel(key, value)
+
+    def _initialize(self, *args):
+        super()._initialize(*args)
+
+        self.output_register = self.base._next_output_register
+        self.input_register = self.base._next_input_register
+
+        self.base._next_output_register = self.output_register + 8
+        self.base._next_input_register = self.input_register + 4
+
+    @CpxBase._require_base
+    def read_channels(self) -> list[int]:
+        """read all channels as a list of integer values"""
+        channel_data = [0, 0, 0, 0]
+        for ch in range(4):
+            raw_data = self.base.read_reg_data(self.input_register + ch, length=4)
+            channel_data[ch] = [CpxBase._decode_int([x]) for x in raw_data]
+        return channel_data
+
+    @CpxBase._require_base
+    def read_status(self) -> list[bool]:
+        """read module status register. Further information see module datasheet"""
+        data = self.base.read_reg_data(self.input_register + 4)[0]
+        return [d == "1" for d in bin(data)[2:].zfill(16)[::-1]]
+
+    @CpxBase._require_base
+    def read_channel(self, channel: int) -> int:
+        """read back the value of one channel"""
+        return self.read_channels()[channel]
+
+    @CpxBase._require_base
+    def read_outputs(self) -> int:
+        """read back the value of one output channel"""
+        raw_data = self.base.read_reg_data(self.input_register + 4, length=4)
+        data = [CpxBase._decode_int([x]) for x in raw_data]
+        return data
+
+    @CpxBase._require_base
+    def read_output_channel(self, channel: int) -> int:
+        """read back the value of one channel"""
+        return self.read_outputs()[channel]
+
+    @CpxBase._require_base
+    def write_channels(self, data: list[int]) -> None:
+        """write data to module channels in ascending order"""
+        reg_data = [self.base._decode_int([x]) for x in data]
+        self.base.write_reg_data(reg_data, self.output_register, length=4)
+
+    @CpxBase._require_base
+    def write_channel(self, channel: int, data: int) -> None:
+        """write data to module channel number"""
+
+        reg_data = self.base._decode_int([data])
+        self.base.write_reg_data(reg_data, self.output_register + channel)
+
+    @CpxBase._require_base
+    def configure_monitoring_uload(self, value: bool) -> None:
+        """The "Monitoring Uload" parameter defines whether the monitoring of the load voltage supply shall be
+        activated or deactivated in regard to undervoltage.
+        When the monitoring is activated, the error is sent to the bus module and indicated by the error LED
+        on the module
+        """
+        function_number = 4828 + 64 * self.position
+        reg = self.base.read_function_number(function_number)
+
+        # Fill in the unchanged values from the register
+        if value:
+            value_to_write = reg | 0x40
+        else:
+            value_to_write = reg & 0xFB
+
+        self.base.write_function_number(function_number, value_to_write)
+
+    @CpxBase._require_base
+    def configure_behaviour_after_scl(self, value: bool) -> None:
+        """The "Behaviour after SCS" parameter defines whether the voltage remains deactivated or reactivates
+        automatically after a short circuit or overload at the IO-Link® interfaces (ports).
+        The voltage can be switched on again with the "leave switched off" setting by deactivating and then
+        reactivating the "PS supply" parameter. Otherwise the activation and deactivation of the
+        automation system CPX-E is required to restore the voltage.
+        """
+        function_number = 4828 + 64 * self.position + 1
+        reg = self.base.read_function_number(function_number)
+
+        # Fill in the unchanged values from the register
+        if value:
+            value_to_write = reg | 0x01
+        else:
+            value_to_write = reg & 0xFE
+
+        self.base.write_function_number(function_number, value_to_write)
+
+    @CpxBase._require_base
+    def configure_behaviour_after_sco(self, value: bool) -> None:
+        """The "Behaviour after SCO" parameter defines whether the voltage remains deactivated or reactivates
+        automatically after a short circuit or overload at the IO-Link® interfaces (ports).
+        The voltage can be switched on again with the "leave switched off" setting by deactivating and then
+        reactivating the "PS supply" (è Tab. 19 ) parameter. Otherwise the activation and deactivation of the
+        automation system CPX-E is required to restore the voltage.
+        """
+        function_number = 4828 + 64 * self.position + 1
+        reg = self.base.read_function_number(function_number)
+
+        # Fill in the unchanged values from the register
+        if value:
+            value_to_write = reg | 0x02
+        else:
+            value_to_write = reg & 0xFD
+
+        self.base.write_function_number(function_number, value_to_write)
+
+    @CpxBase._require_base
+    def configure_ps_supply(self, value: bool) -> None:
+        """The "PS supply" parameter defines whether the operating voltage supply shall be deactivated or activated.
+        The setting applies for all IO-Link interfaces (ports).
+        """
+        function_number = 4828 + 64 * self.position + 6
+        reg = self.base.read_function_number(function_number)
+
+        # Fill in the unchanged values from the register
+        if value:
+            value_to_write = reg | 0x01
+        else:
+            value_to_write = reg & 0xFE
+
+        self.base.write_function_number(function_number, value_to_write)
+
+    @CpxBase._require_base
+    def configure_cycle_time(
+        self, value: tuple[int], channel: int | list | None = None
+    ) -> None:
+        """The "Cycle time" parameter defines the cycle time (low/high) set by the IO-Link master.
+        The setting can be made separately for each IO-Link interface (port).
+        The value becomes effective at the start of the IO-Link connection by setting the "Operating mode"
+        parameter to "IO-Link". Changes during IO-Link operation are not made until the connection
+        has been deactivated and then reactivated again.
+        Values are 16 bit in us unit. Default is 0 (minimum supported cycle time) for both low and high.
+        If no channels are specified, all channels are set to the given value.
+        """
+        if channel is None:
+            channel = [0, 1, 2, 3]
+
+        function_number = [
+            4828 + 64 * self.position + 8,
+            4828 + 64 * self.position + 12,
+            4828 + 64 * self.position + 16,
+            4828 + 64 * self.position + 20,
+        ]
+
+        if isinstance(channel, int):
+            channel = [channel]
+
+        for ch in channel:
+            self.base.write_function_number(function_number[ch], value[0])
+            self.base.write_function_number(function_number[ch] + 1, value[1])
+
+    @CpxBase._require_base
+    def configure_pl_supply(
+        self, value: bool, channel: int | list | None = None
+    ) -> None:
+        """The "PL supply" parameter defines whether the load voltage supply shall be deactivated or activated.
+        The setting can be made separately for each IO-Link interface (port). If no channel is specified,
+        the value will be applied to all channels.
+        """
+        if channel is None:
+            channel = [0, 1, 2, 3]
+
+        function_number = [
+            4828 + 64 * self.position + 10,
+            4828 + 64 * self.position + 14,
+            4828 + 64 * self.position + 18,
+            4828 + 64 * self.position + 22,
+        ]
+
+        if isinstance(channel, int):
+            channel = [channel]
+
+        for ch in channel:
+            reg = self.base.read_function_number(function_number[ch])
+
+            # Fill in the unchanged values from the register
+            if value:
+                value_to_write = reg | 0x01
+            else:
+                value_to_write = reg & 0xFE
+
+            self.base.write_function_number(function_number[ch], value_to_write)
+
+    @CpxBase._require_base
+    def configure_operating_mode(
+        self, value: int, channel: int | list | None = None
+    ) -> None:
+        """The "Operating mode" parameter defines the operating mode of the IO-Link® interface (port).
+        The setting can be made separately for each IO-Link interface (port).
+        Possible Values are:
+        - 0: Inactive: Port is not in use (default)
+        - 1: DI: Port acts like a digital input
+        - 2: [DO]: reserved
+        - 3: IO-Link: IO-Link communication
+        """
+        if channel is None:
+            channel = [0, 1, 2, 3]
+
+        function_number = [
+            4828 + 64 * self.position + 11,
+            4828 + 64 * self.position + 15,
+            4828 + 64 * self.position + 19,
+            4828 + 64 * self.position + 23,
+        ]
+
+        if value not in range(4):
+            raise ValueError("Operating mode must be between 0 and 3")
+
+        if isinstance(channel, int):
+            channel = [channel]
+
+        for ch in channel:
+            # delete two lsb from register to write the new value there
+            reg = self.base.read_function_number(function_number[ch]) & 0xFC
+
+            value_to_write = reg | value
+
+            self.base.write_function_number(function_number[ch], value_to_write)
+
+    @CpxBase._require_base
+    def read_line_state(self, channel: int | list | None = None) -> list[str] | str:
+        """Line state for all channels. If no channel is provided, list of all channels is returned."""
+        if channel is None:
+            channel = [0, 1, 2, 3]
+
+        function_number = [
+            4828 + 64 * self.position + 24,
+            4828 + 64 * self.position + 27,
+            4828 + 64 * self.position + 30,
+            4828 + 64 * self.position + 33,
+        ]
+
+        line_state = []
+        for ch in range(4):
+            reg = self.base.read_function_number(function_number[ch]) & 0x07
+            if reg == 0b000:
+                state = "Inactive"
+            elif reg == 0b001:
+                state = "DI"
+            elif reg == 0b011:
+                state = "CheckFault"
+            elif reg == 0b100:
+                state = "Preoperate"
+            elif reg == 0b101:
+                state = "Operate"
+            elif reg == 0b110:
+                state = "Scanning"
+            elif reg == 0b111:
+                state = "DeviceLost"
+            else:
+                raise ValueError(f"Read unknown linestate {reg} for channel {ch}")
+            line_state.append(state)
+
+        if isinstance(channel, int):
+            return line_state[channel]
+
+        return line_state
+
+    @CpxBase._require_base
+    def read_device_error(self, channel: int | list | None = None) -> tuple[int] | int:
+        """the "Device error code" parameter displays the current lowest-value error code (event code) of the
+        connected IO-Link device. If no event is reported, the parameter has a value of 0.
+        Returns list of tuples of (Low, High) values in hexadecimal strings for each requested channel. If only one channel
+        is requested, only one tuple is returned.
+        """
+        if channel is None:
+            channel = [0, 1, 2, 3]
+
+        function_number = [
+            4828 + 64 * self.position + 25,
+            4828 + 64 * self.position + 28,
+            4828 + 64 * self.position + 31,
+            4828 + 64 * self.position + 34,
+        ]
+
+        device_error = [0, 0, 0, 0]
+        for ch in range(4):
+            low = self.base.read_function_number(function_number[ch])
+            high = self.base.read_function_number(function_number[ch] + 1)
+            device_error[ch] = (hex(low), hex(high))
+
+        if isinstance(channel, int):
+            return device_error[channel]
+
+        return device_error
 
 
-class CpxE1Ci(_CpxEModule):
+class CpxE1Ci(CpxEModule):
     """Class for CPX-E-1CI counter module"""
 
     # TODO: Add 1Cl module